requirements.txt
filetype==1.0.6
lxml==4.5.0
python-dateutil==2.8.1

# Development dependencies
autopep8==1.5.1
pytest==5.4.1